refactor: :truck: move organizations route
