+++ apps/client/src/stores/admin/organization.ts
@@ -3,40 +3,39 @@ import {
   type Organization,
   type CreateOrganizationBody,
   type UpdateOrganizationBody,
-  type UpdateOrganizationParams,
   resourceListToDict,
 } from '@cpn-console/shared'
 import { apiClient, extractData } from '@/api/xhr-client.js'
 
-export const useAdminOrganizationStore = defineStore('admin-organization', () => {
+export const useOrganizationStore = defineStore('organization', () => {
   const organizations = ref<Organization[]>([])
   const organizationsById = computed(() => resourceListToDict(organizations.value))
 
-  const getAllOrganizations = async () => {
-    organizations.value = await apiClient.OrganizationsAdmin.getAllOrganizations()
+  const listOrganizations = async () => {
+    organizations.value = await apiClient.Organizations.listOrganizations()
       .then(response => extractData(response, 200))
     return organizations.value
   }
 
   const createOrganization = (organization: CreateOrganizationBody) =>
-    apiClient.OrganizationsAdmin.createOrganization({ body: { ...organization, source: 'dso-console' } })
+    apiClient.Organizations.createOrganization({ body: { ...organization, source: 'dso-console' } })
       .then(response => extractData(response, 201))
 
-  const updateOrganization = (organization: UpdateOrganizationBody & { name: UpdateOrganizationParams['organizationName'] }) =>
-    apiClient.OrganizationsAdmin.updateOrganization({
+  const updateOrganization = (organization: UpdateOrganizationBody & { name: Organization['name'] }) =>
+    apiClient.Organizations.updateOrganization({
       body: { ...organization, source: 'dso-console' },
       params: { organizationName: organization.name },
     })
       .then(response => extractData(response, 200))
 
   const fetchOrganizations = () =>
-    apiClient.OrganizationsAdmin.syncOrganizations()
+    apiClient.Organizations.syncOrganizations()
       .then(response => extractData(response, 200))
 
   return {
     organizations,
     organizationsById,
-    getAllOrganizations,
+    listOrganizations,
     createOrganization,
     updateOrganization,
     fetchOrganizations,
