+++ apps/client/src/stores/organization.ts
@@ -1,7 +1,9 @@
 import { defineStore } from 'pinia'
-import { ref } from 'vue'
+import { computed, ref } from 'vue'
 import {
   type Organization,
+  type CreateOrganizationBody,
+  type UpdateOrganizationBody,
   resourceListToDict,
 } from '@cpn-console/shared'
 import { apiClient, extractData } from '@/api/xhr-client.js'
@@ -10,15 +12,33 @@ export const useOrganizationStore = defineStore('organization', () => {
   const organizations = ref<Organization[]>([])
   const organizationsById = computed(() => resourceListToDict(organizations.value))
 
-  const setOrganizations = async () => {
-    organizations.value = await apiClient.Organizations.getOrganizations()
+  const listOrganizations = async () => {
+    organizations.value = await apiClient.Organizations.listOrganizations()
       .then(response => extractData(response, 200))
     return organizations.value
   }
 
+  const createOrganization = (organization: CreateOrganizationBody) =>
+    apiClient.Organizations.createOrganization({ body: { ...organization, source: 'dso-console' } })
+      .then(response => extractData(response, 201))
+
+  const updateOrganization = (organization: UpdateOrganizationBody & { name: Organization['name'] }) =>
+    apiClient.Organizations.updateOrganization({
+      body: { ...organization, source: 'dso-console' },
+      params: { organizationName: organization.name },
+    })
+      .then(response => extractData(response, 200))
+
+  const syncOrganizations = () =>
+    apiClient.Organizations.syncOrganizations()
+      .then(response => extractData(response, 200))
+
   return {
     organizations,
     organizationsById,
-    setOrganizations,
+    listOrganizations,
+    createOrganization,
+    updateOrganization,
+    syncOrganizations,
   }
 })
