+++ apps/client/src/views/admin/ListClusters.vue
@@ -3,7 +3,7 @@ import { ref, computed, onMounted, watch } from 'vue'
 import { sortArrByObjKeyAsc, type CreateClusterBody, type UpdateClusterBody, type Cluster, type Stage, type Project, type ClusterAssociatedEnvironments } from '@cpn-console/shared'
 import { useAdminClusterStore } from '@/stores/admin/cluster.js'
 import { useAdminProjectStore } from '@/stores/admin/project.js'
-import { useAdminOrganizationStore } from '@/stores/admin/organization.js'
+import { useOrganizationStore } from '@/stores/organization.js'
 import { useZoneStore } from '@/stores/zone.js'
 import { useStageStore } from '@/stores/stage.js'
 import { useClusterStore } from '@/stores/cluster'
@@ -17,7 +17,7 @@ type ClusterList = {
 const adminClusterStore = useAdminClusterStore()
 const clusterStore = useClusterStore()
 const adminProjectStore = useAdminProjectStore()
-const adminOrganizationStore = useAdminOrganizationStore()
+const organizationStore = useOrganizationStore()
 const zoneStore = useZoneStore()
 const stageStore = useStageStore()
 
@@ -97,7 +97,7 @@ onMounted(async () => {
   setClusterTiles(clusters.value)
   await zoneStore.getAllZones()
   allProjects.value = await adminProjectStore.getAllActiveProjects()
-  const organizations = await adminOrganizationStore.getAllOrganizations()
+  const organizations = await organizationStore.listOrganizations()
   allProjects.value.forEach(project => { project.organization = organizations.find(org => org.id === project.organizationId) })
   allStages.value = await stageStore.getAllStages()
 })
