+++ apps/server/src/resources/index-admin.ts
@@ -1,7 +1,6 @@
 import { type FastifyInstance } from 'fastify'
 import { clusterAdminRouter } from './cluster/admin/router.js'
 import { logAdminRouter } from './log/admin/router.js'
-import { organizationAdminRouter } from './organization/admin/router.js'
 import { projectAdminRouter } from './project/admin/router.js'
 import { quotaAdminRouter } from './quota/admin/router.js'
 import { stageAdminRouter } from './stage/admin/router.js'
@@ -14,7 +13,6 @@ export const apiRouterAdmin = () => async (app: FastifyInstance) => {
   app.addHook('preHandler', checkAdminGroup)
   await app.register(serverInstance.plugin(clusterAdminRouter()), { responseValidation: true })
   await app.register(serverInstance.plugin(logAdminRouter()))
-  await app.register(serverInstance.plugin(organizationAdminRouter()))
   await app.register(serverInstance.plugin(projectAdminRouter()))
   await app.register(serverInstance.plugin(quotaAdminRouter()))
   await app.register(serverInstance.plugin(stageAdminRouter()))
