+++ packages/shared/package.json
@@ -33,7 +33,7 @@
     "@vitest/coverage-v8": "^1.6.0",
     "nodemon": "^3.1.4",
     "rimraf": "^5.0.7",
-    "typescript": "^5.5.2",
+    "typescript": "5.4.5",
     "vite": "^5.3.2",
     "vitest": "^1.6.0"
   },
