+++ packages/shared/src/api-client.ts
@@ -21,7 +21,6 @@ export const getContract = async () => contractInstance.router(
     Zones: (await import('./contracts/index.js')).zoneContract,
 
     ClustersAdmin: (await import('./contracts/index.js')).clusterAdminContract,
-    OrganizationsAdmin: (await import('./contracts/index.js')).organizationAdminContract,
     ProjectsAdmin: (await import('./contracts/index.js')).projectAdminContract,
     QuotasAdmin: (await import('./contracts/index.js')).quotaAdminContract,
     StagesAdmin: (await import('./contracts/index.js')).stageAdminContract,
