+++ packages/shared/src/utils/schemas.ts
@@ -1,5 +1,5 @@
 // import Joi from 'joi'
-import type { ZodError, SafeParseReturnType, ZodObject } from 'zod'
+import { type ZodError, type SafeParseReturnType, type ZodObject, z } from 'zod'
 import { fromZodError } from 'zod-validation-error'
 
 export type SharedZodError = ZodError
@@ -16,3 +16,7 @@ export const instanciateSchema = (schema: ZodObject<any>, value: unknown) => {
   }
   return {}
 }
+
+export const CoerceBooleanSchema = z.boolean()
+  .or(z.enum(['true', 'false'])
+    .transform((value) => value === 'true'))
